+++ elements/drawtools/src/enums/draw.js
@@ -1,0 +1,1 @@
+export const FIT_OPTIONS = { duration: 750 };