+++ elements/drawtools/src/enums/test.js
@@ -1,0 +1,6 @@
+export const TEST_SELECTORS = {
+  drawTools: "eox-drawtools",
+  controller: "eox-drawtools-controller",
+  drawBtn: "[data-cy='drawBtn']",
+  discardBtn: "[data-cy='discardBtn']",
+};