+++ elements/drawtools/stories/index.js
@@ -1,0 +1,7 @@
+// Export different stories
+
+export { default as PrimaryStory } from "./primary"; // Primary story showcasing basic usage.
+export { default as DrawTypeStory } from "./draw-type"; // Story demonstrating multiple polygons.
+export { default as MultiPolygonStory } from "./multi-polygon"; // Story showcasing feature modification.
+export { default as ModifyFeaturesStory } from "./modify-features"; // Story illustrating different draw types.
+export { default as MultiPolygonWithListStory } from "./multi-polygon-list"; // Story demonstrating polygons with lists.