+++ elements/drawtools/stories/multi-polygon-list.js
@@ -1,0 +1,32 @@
+/**
+ * Component demonstrating the display of a list of features on the eox-map.
+ * This component showcases the usage of eox-drawtools with the `show-list` attribute/property set
+ * to display a list of features corresponding to drawn polygons on the map.
+ */
+import { html } from "lit";
+import "../../map/main";
+import "../src/main";
+import { STORIES_LAYERS_ARRAY, STORIES_MAP_STYLE } from "../src/enums";
+
+export const MultiPolygonWithList = {
+  render: () => html`
+    <!-- Render eox-map component with ID "list" -->
+    <div style="display: flex">
+      <eox-map
+        id="list"
+        style=${STORIES_MAP_STYLE}
+        .layers=${STORIES_LAYERS_ARRAY}
+      ></eox-map>
+
+      <!-- Initialize eox-drawtools for the eox-map with ID "list" -->
+      <eox-drawtools
+        for="eox-map#list"
+        layer="draw"
+        multiple-features
+        show-list
+      ></eox-drawtools>
+    </div>
+  `,
+};
+
+export default MultiPolygonWithList;