+++ elements/drawtools/test/cases/index.js
@@ -1,0 +1,5 @@
+// Exported test methods
+
+export { default as clickDiscardBtnTest } from "./click-discard-btn"; // Test for clicking the discard button
+export { default as clickDrawBtnTest } from "./click-draw-btn"; // Test for clicking the draw button
+export { default as loadDrawToolsTest } from "./load-drawtool"; // Test for loading the drawtools component