+++ elements/itemfilter/src/filters/multiselect.ts
@@ -58,7 +58,9 @@ export class EOxItemFilterMultiselect extends LitElement {
     this.filterObject.stringifiedState = Object.keys(this.filterObject.state)
       .filter((k) => this.filterObject.state[k])
       .join(", ");
-    this.filterObject.dirty = true;
+    if (this.filterObject.stringifiedState?.length > 0) {
+      this.filterObject.dirty = true;
+    }
     this.dispatchEvent(new CustomEvent("filter"));
   }